\name{checkDuplicates}
\alias{checkDuplicates}
\title{Checks the feature data set for duplicated instances.}
\usage{
checkDuplicates(astask)
}
\arguments{
  \item{astask}{[\code{\link{ASTask}}]\cr Algorithm
  selection task.}
}
\value{
Printed output of blocks of duplicated features (if
existing).
}
\description{
Checks the feature data set for duplicated instances.
}

