\name{findDominatedAlgos}
\alias{findDominatedAlgos}
\title{Creates a table that shows the dominance of one algorithm over another one.}
\usage{
findDominatedAlgos(astask)
}
\arguments{
  \item{astask}{[\code{\link{ASTask}}]\cr Algorithm
  selection task.}
}
\value{
[\code{data.frame}].  Overview of the dominance between the
algorithms.
}
\description{
Creates a table that shows the dominance of one algorithm
over another one.
}

