\name{getAlgorithmNames}
\alias{getAlgorithmNames}
\title{Returns algorithm names of scenario.}
\usage{
getAlgorithmNames(asscenario)
}
\arguments{
  \item{asscenario}{[\code{\link{ASScenario}}]\cr Algorithm
  selection scenario.}
}
\value{
[\code{character}].
}
\description{
Returns algorithm names of scenario.
}

