\name{getDefaultFeatureStepNames}
\alias{getDefaultFeatureStepNames}
\title{Returns the default feature step names of scenario.}
\usage{
getDefaultFeatureStepNames(asscenario)
}
\arguments{
  \item{asscenario}{[\code{\link{ASScenario}}]\cr Algorithm
  selection scenario.}
}
\value{
[\code{character}].
}
\description{
Returns the default feature step names of scenario.
}

