\name{getFeatureNames}
\alias{getFeatureNames}
\title{Returns feature names of scenario.}
\usage{
getFeatureNames(asscenario)
}
\arguments{
  \item{asscenario}{[\code{\link{ASScenario}}]\cr Algorithm
  selection scenario.}
}
\value{
[\code{character}].
}
\description{
Returns feature names of scenario.
}

