\name{getFeatureStepNames}
\alias{getFeatureStepNames}
\title{Returns feature step names of scenario.}
\usage{
getFeatureStepNames(asscenario)
}
\arguments{
  \item{asscenario}{[\code{\link{ASScenario}}]\cr Algorithm
  selection scenario.}
}
\value{
[\code{character}].
}
\description{
Returns feature step names of scenario.
}

