\name{getInstanceNames}
\alias{getInstanceNames}
\title{Returns instance names of scenario.}
\usage{
getInstanceNames(asscenario)
}
\arguments{
  \item{asscenario}{[\code{\link{ASScenario}}]\cr Algorithm
  selection scenario.}
}
\value{
[\code{character}].
}
\description{
Returns instance names of scenario.
}

