\name{getNumberOfCVFolds}
\alias{getNumberOfCVFolds}
\title{Returns number of CV folds.}
\usage{
getNumberOfCVFolds(asscenario)
}
\arguments{
  \item{asscenario}{[\code{\link{ASScenario}}]\cr Algorithm
  selection scenario.}
}
\value{
[\code{integer(1)}].
}
\description{
Returns number of CV folds.
}

