\name{getNumberOfCVReps}
\alias{getNumberOfCVReps}
\title{Returns number of CV repetitions.}
\usage{
  getNumberOfCVReps(astask)
}
\arguments{
  \item{astask}{[\code{\link{ASTask}}]\cr Algorithm
  selection task.}
}
\value{
  [\code{integer(1)}].
}
\description{
  Returns number of CV repetitions.
}

