\name{getProvidedFeatures}
\alias{getProvidedFeatures}
\title{Return features that are useable for a given set of feature steps.}
\usage{
getProvidedFeatures(astask, steps)
}
\arguments{
  \item{astask}{[\code{\link{ASTask}}]\cr Algorithm
  selection task.}

  \item{steps}{[\code{character}]\cr Feature steps.
  Default are all feature steps.}
}
\value{
[\code{character}].
}
\description{
Return features that are useable for a given set of feature
steps.
}

