\name{lookForDuplicates}
\alias{lookForDuplicates}
\title{Checks the feature data set for duplicated instances.}
\usage{
lookForDuplicates(astask)
}
\arguments{
  \item{astask}{[\code{\link{ASTask}}]\cr Algorithm
  selection task.}
}
\value{
Blocks of duplicated features (if existing).
}
\description{
Checks the feature data set for duplicated instances.
}

