\name{parseASTask}
\alias{ASTask}
\alias{parseASTask}
\title{Parses description file and returns an S3 class of the contents}
\usage{
  parseASTask(path)
}
\arguments{
  \item{path}{[\code{character(1)}]\cr Path to directory of
  benchmark data set.}
}
\value{
  [\code{\link{ASTask}}]. Description object.
}
\description{
  Parses description file and returns an S3 class of the
  contents
}

