\name{summarizeAlgoRunstatus}
\alias{summarizeAlgoRunstatus}
\title{Creates a table that summarizes the runstatus of each algorithm.}
\usage{
summarizeAlgoRunstatus(astask)
}
\arguments{
  \item{astask}{[\code{\link{ASTask}}]\cr Algorithm
  selection task.}
}
\value{
[\code{table}].  Table, which summarizes the runstatus per
algorithm.
}
\description{
Creates a table that summarizes the runstatus of each
algorithm.
}

