\name{summarizeFeatureSteps}
\alias{summarizeFeatureSteps}
\title{Creates a data.frame that summarizes the feature steps.}
\usage{
  summarizeFeatureSteps(astask)
}
\arguments{
  \item{astask}{[\code{\link{ASTask}}]\cr Algorithm
  selection task.}
}
\value{
  [\code{data.frame}].
}
\description{
  Creates a data.frame that summarizes the feature steps.
}

