\name{summarizeFeatureValues}
\alias{summarizeFeatureValues}
\title{Creates summary data.frame for feature values across all instances.}
\usage{
summarizeFeatureValues(astask)
}
\arguments{
  \item{astask}{[\code{\link{ASTask}}]\cr Algorithm
  selection task.}
}
\value{
[\code{data.frame}].
}
\description{
Creates summary data.frame for feature values across all
instances.
}

