% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getAlgorithmNames}
\alias{getAlgorithmNames}
\title{Returns algorithm names of scenario.}
\usage{
getAlgorithmNames(asscenario)
}
\arguments{
\item{asscenario}{[\code{\link{ASScenario}}]\cr
Algorithm selection scenario.}
}
\value{
[\code{character}].
}
\description{
Returns algorithm names of scenario.
}

