% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getCostsAndPresolvedStatus}
\alias{getCostsAndPresolvedStatus}
\title{Return wether an instance was presolved and which step did it.}
\usage{
getCostsAndPresolvedStatus(asscenario, feature.steps)
}
\arguments{
\item{asscenario}{[\code{\link{ASScenario}}]\cr
Algorithm selection scenario.}

\item{feature.steps}{[\code{character}]\cr
Which feature steps are allowed?
Default is all steps.}
}
\value{
[\code{list}]. In the following, \code{n} is the number of instances. All following object are ordered by \dQuote{instance_id}.
  \item{is.presolved [\code{logical(n)}]}{Was instance presolved? Named by instance ids.}
  \item{solve.steps [\code{character(n)}]}{Which step solved it? NA if no step did it. Named by instance ids.}
  \item{costs [\code{numeric(n)}]}{Feature costs for using the steps. Named by instance ids. NULL if no costs are present.}
}
\description{
Return wether an instance was presolved and which step did it.
}

