\name{getDefaultFeatureStepNames}
\alias{getDefaultFeatureStepNames}
\title{Returns the default feature step names of task.}
\usage{
  getDefaultFeatureStepNames(astask)
}
\arguments{
  \item{astask}{[\code{\link{ASTask}}]\cr Algorithm
  selection task.}
}
\value{
  [\code{character}].
}
\description{
  Returns the default feature step names of task.
}

