\name{getFeatureStepNames}
\alias{getFeatureStepNames}
\title{Returns feature step names of task.}
\usage{
  getFeatureStepNames(astask)
}
\arguments{
  \item{astask}{[\code{\link{ASTask}}]\cr Algorithm
  selection task.}
}
\value{
  [\code{character}].
}
\description{
  Returns feature step names of task.
}

