\name{getNumberOfCVReps}
\alias{getNumberOfCVReps}
\title{Returns number of CV repetitions.}
\usage{
getNumberOfCVReps(asscenario)
}
\arguments{
  \item{asscenario}{[\code{\link{ASScenario}}]\cr Algorithm
  selection scenario.}
}
\value{
[\code{integer(1)}].
}
\description{
Returns number of CV repetitions.
}

