\name{getSummedFeatureCosts}
\alias{getSummedFeatureCosts}
\title{Returns feature costs of scenario, summed over all instances.}
\usage{
getSummedFeatureCosts(asscenario, feature.steps)
}
\arguments{
  \item{asscenario}{[\code{\link{ASScenario}}]\cr Algorithm
  selection scenario.}

  \item{feature.steps}{[\code{character}]\cr Sum costs only
  for these selected steps.  Default are all feature
  steps.}
}
\value{
[\code{character}].
}
\description{
Returns feature costs of scenario, summed over all
instances.
}

