\name{getSummedFeatureCosts}
\alias{getSummedFeatureCosts}
\title{Returns feature costs of task, summed over all instances.}
\usage{
  getSummedFeatureCosts(astask, feature.steps)
}
\arguments{
  \item{astask}{[\code{\link{ASTask}}]\cr Algorithm
  selection task.}

  \item{feature.steps}{[\code{character}]\cr Sum costs only
  for these selected steps.  Default are all feature
  steps.}
}
\value{
  [\code{character}].
}
\description{
  Returns feature costs of task, summed over all instances.
}

