\name{summarizeAlgoRunstatus}
\alias{summarizeAlgoRunstatus}
\title{Creates summary data.frame for algorithm runstatus across all instances.}
\usage{
summarizeAlgoRunstatus(astask)
}
\arguments{
  \item{astask}{[\code{\link{ASTask}}]\cr Algorithm
  selection task.}
}
\value{
[\code{data.frame}].
}
\description{
Creates summary data.frame for algorithm runstatus across
all instances.
}

