% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summarizeAlgoRunstatus}
\alias{summarizeAlgoRunstatus}
\title{Creates summary data.frame for algorithm runstatus across all instances.}
\usage{
summarizeAlgoRunstatus(asscenario)
}
\arguments{
\item{asscenario}{[\code{\link{ASScenario}}]\cr
Algorithm selection scenario.}
}
\value{
[\code{data.frame}].
}
\description{
Creates summary data.frame for algorithm runstatus across all instances.
}

