\name{summarizeFeatureRunStatus}
\alias{summarizeFeatureRunStatus}
\title{Creates a table that summarize the runstatus of the features.}
\usage{
summarizeFeatureRunStatus(astask)
}
\arguments{
  \item{astask}{[\code{\link{ASTask}}]\cr Algorithm
  selection task.}
}
\value{
[\code{table}].  Table, which summarizes the runstatus per
algorithm.
}
\description{
Creates a table that summarize the runstatus of the
features.
}

