\name{summarizeFeatureSteps}
\alias{summarizeFeatureSteps}
\title{Creates a data.frame that summarizes the feature steps.}
\usage{
summarizeFeatureSteps(asscenario)
}
\arguments{
  \item{asscenario}{[\code{\link{ASScenario}}]\cr Algorithm
  selection scenario.}
}
\value{
[\code{data.frame}].
}
\description{
Creates a data.frame that summarizes the feature steps.
}

