\name{summarizeFeatureValues}
\alias{summarizeFeatureValues}
\title{Creates a table that gives an overview of the feature values
across all instances.}
\usage{
  summarizeFeatureValues(astask)
}
\arguments{
  \item{astask}{[\code{\link{ASTask}}]\cr Algorithm
  selection task.}
}
\value{
  [\code{data.frame}].  Overview of the feature values.
}
\description{
  Creates a table that gives an overview of the feature
  values across all instances.
}

